Add simple TrayerWidget (#6)
This commit adds a new TrayerWidget to incorporate the system tray
utility 'trayer' into barpyrus. As of now it supports all screen corners
as positions on the bar, any other position might need further work.

The TrayerWidget itself instantiates a TrayerWatch object to start the
trayer process and a new X connection. It is then used to passively
watch for any updates of the trayer regarding its width using already
existing mechanisms of barpyrus.