Fix racecondition in TrayerWidget
As of now Barpyrus might stop working due to the trayer process not
spawning its window before TrayerWatch starts looking for it. In order
to fix this issue, TrayerWatch now passively waits for any window change
and then checks for the existence of a trayer window until it finds it.