Add ConkyGenerator.match()
This adds a simplified syntax for conditionals based on conky variables.

Instead of:

    with cg.if_('match ${ibm_fan} != 65535'):
        ...

it's now possible to write

    with cg.if_(cg.match('ibm_fan') != 65535):
        ...

i.e. using Python types (str/int/float) and comparison operators
(>,>=,<,<=,!=,==). This covers all usages of `if_match` permitted by conky.

This is especially noticable when string formatting is involved, as conky's
`${...}` interfers with Python f-strings. This example using old-style string
interpolation:

    with cg.if_('match "${addr %s}" != "No Address"' % iface):
        ...

turns into:

    with cg.if_(cg.match(f'addr {iface}') != "No Address"):
        ...

Note that in this example, the LHS is actually a string, and we need to tell
conky about that by quoting the interpolating value. With the new syntax, we
assume the LHS needs quotes if the RHS is a string - after all, conky rejects
comparisons between different types anyways.