Update tag status when a tag was renamed (#13)
Before, renaming a tag wouldn't display the new tag names. As a workaround, `hc
emit_hook tag_flags` can be used after renaming a tag, but it's a bit of a
hack.

This also adds a for loop (in a separate commit) to reduce duplication a bit
and removes the unnecessary lambdas.

Also see https://github.com/herbstluftwm/herbstluftwm/issues/999
- but given that it doesn't use the hook's arguments, any change there should
probably not matter much.